Split Expression into two subtype files
Added AtomicExpression and CompoundExpression;
Moved some classes into ExpressionFunction.
Expression is no longer a sealed trait but other seals are intact.
ExpressionMonoFunction and ExpressionBiFunction are now sealed.
